moved relatedrecord proxy and get_context to thread
