+++ llm_tool/models/__init__.py
@@ -1,6 +1,6 @@
 from . import llm_tool
 from . import llm_tool_record_retriever
-from . import llm_thread
+from . import llm_provider
 from . import llm_model
 from . import mail_message
 from . import llm_tool_consent_config
